requirements.txt
asgiref==3.11.0
Django==5.2.10
django-filter==25.2
django-markdownx==4.0.9
djangorestframework==3.16.1
Markdown==3.10.1
pillow==12.1.0
sqlparse==0.5.5
tzdata==2025.3
